requirements.txt
fastapi
openai==1.63.0
torch==2.5.0
onnxruntime
pyannote.audio==3.3.2
uvicorn
tempfile
pydub
python-multipart==0.0.20
python-dotenv